Moved middleware into `middleware folder` and updated rferences
