+++ backend/package.json
@@ -25,6 +25,7 @@
 		"debug": "~2.6.9",
 		"dotenv": "^16.4.1",
 		"express": "^4.18.2",
+		"express-validator": "^7.0.1",
 		"http-errors": "~1.6.3",
 		"jsonwebtoken": "^9.0.2",
 		"morgan": "~1.9.1",
